I have made some changes in code: 1. I had problem with running ElasticSearch part on my computer: - kafka-elasticsearch part was not identified as module - ElasticSearchAllData.java: main() function was not runnable - ElasticSearchFilterData.java: main() function was not runnable That's why - I created new module with same name - kafka-elasticsearch - I have moved ElasticSearchAllData.java and ElasticSearchFilterData.java into new module's java folder 2. I have added intructions.sh for kafka-elasticsearch module too 3. Also changed kafka-connect/AnomalyDetection.ipynb: now it saves balance.csv in data folder, so no need to move by hand I have checked workflow it is working
