+++ kafka-connect/balance.properties
@@ -10,6 +10,7 @@ error.path=/Users/joe/Documents/ANDROID/idea/AnomalyDetection/kafka-connect/erro
 halt.on.error=false
 topic=balance_topic
 csv.first.row.as.header=true
+batch.size = 1000
 
 key.schema={"name":"com.elte.anomalydetection.BalanceKey","type":"STRUCT","isOptional":false,"fieldSchemas":{"id":{"type":"INT64","isOptional":false}}}
 value.schema={"name":"com.elte.anomalydetection.Balance","type":"STRUCT","isOptional":false,"fieldSchemas":{"id":{"type":"INT64","isOptional":false},"EBIZ_BALANCE_ID":{"type": "STRING", "isOptional": true},"VALID_FROM":{"type": "STRING", "isOptional": true},"VALID_TO":{"type": "STRING", "isOptional": true},"GIRONUMBER":{"type": "STRING", "isOptional": true},"AMOUNT":{"type": "STRING", "isOptional": true},"CURRENCY":{"type": "STRING", "isOptional": true}}}
